+++ docs/changelog.md
@@ -42,6 +42,10 @@
 
 ## 更新说明
 
+1.8.20 fix: fix stock_board_concept_hist_em interface
+
+    1. 修复 stock_board_concept_hist_em 接口
+
 1.8.19 fix: fix stock_balance_sheet_by_yearly_em interface
 
     1. 修复 stock_balance_sheet_by_yearly_em 接口
@@ -1605,6 +1609,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.8.20 fix: fix stock_board_concept_hist_em interface
 
 1.8.19 fix: fix stock_balance_sheet_by_yearly_em interface